Add pingInterval option - Fix #554
The server is closing connection after a time, which makes can cause
errors in several environments.

This commit adds a new config option when initializing thinky, which
will help keeping the connection alive, using `rethinkdbdash`'s solution
with the same input.

- `pingInterval`: - if `> 0`, the connection will be pinged every
  `pingInterval` seconds, default `-1`
  https://github.com/neumino/rethinkdbdash

It was implemented following this discussion, which describes the issue
pretty well:
https://github.com/neumino/rethinkdbdash/issues/192#issuecomment-212753388